Remove mogenius from deployment guide
Hi there, mogenius doesn't have free cloud resources anymore. It is still possible to deploy containers for free but it requires a local Kubernetes. I suppose this is not a use case for this app so I suggest to remove mogenius from the deployment guides.
I'd also suggest to edit the Wiki accordingly.